+++ server/my_db_handler/db_handler.py
@@ -1,13 +1,12 @@
 import firebase_admin
 from firebase_admin import credentials
 from firebase_admin import firestore
-from account import Account
-from account import Account
+from server.entity_classes.account import Account
 
 
 class DatabaseHandler:
-    def __init__(self):
-        self.cred = credentials.Certificate("firebase_credentials/hacked24-60c88-firebase-adminsdk-5fu9m-0ba7ceb240.json")
+    def __init__(self, certificate_json_path):
+        self.cred = credentials.Certificate(certificate_json_path)
         self.app = firebase_admin.initialize_app(self.cred)
         self.db = firestore.client()
     
@@ -29,11 +28,6 @@ class DatabaseHandler:
         """
         docs = self.user_collection().stream()
         return [Account.from_dict(x.to_dict()) for x in docs]
-       
-
-    def save_to_db(self, db):
-        doc_ref = db.collection("users").document(self.username)
-        doc_ref.set(self.to_dict())
 
 
 if __name__ == "__main__":
