+++ server/tests/account_test.py
@@ -1,0 +1,13 @@
+import unittest
+from server.entity_classes.account import Account
+
+
+class MyTestCase(unittest.TestCase):
+
+    def test_account_construction(self):
+        expected = Account("testing1")
+        test_input = {"username": "testing1", "time_spent": 0, "points": 0}
+        self.assertEqual(expected, Account.from_dict(test_input))
+
+if __name__ == '__main__':
+    unittest.main()
